refactor: 오타수정
