+++ src/main/java/vision/cotegory/security/filter/RefreshTokenFilter.java
@@ -6,7 +6,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.http.MediaType;
 import org.springframework.web.filter.OncePerRequestFilter;
-import vision.cotegory.security.exceptoin.RefreshTokenException;
+import vision.cotegory.security.exception.RefreshTokenException;
 import vision.cotegory.security.JWTUtil;
 
 import javax.servlet.FilterChain;
