+++ src/main/java/vision/cotegory/security/filter/TokenCheckFilter.java
@@ -9,7 +9,7 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.filter.OncePerRequestFilter;
-import vision.cotegory.security.exceptoin.AccessTokenException;
+import vision.cotegory.security.exception.AccessTokenException;
 import vision.cotegory.security.APIUserDetailsService;
 import vision.cotegory.security.JWTUtil;
